refacor: test class 나누기
